Commit #5, also boolean chunk reader
